refactor: CustomOAuth2User로 변경 및 디렉토리 구조 변경
- DefaultOAuth2User 대신 userId를 포함한 CustomOAuth2User를 구현하여 사용
- OAuth 관련 파일을 oauth 하위 디렉토리로 이동
