+++ src/main/java/com/moonbaar/common/oauth/CustomOAuth2User.java
@@ -1,0 +1,34 @@
+package com.moonbaar.common.oauth;
+
+import java.util.Collection;
+import java.util.Map;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+@Getter
+@AllArgsConstructor
+public class CustomOAuth2User implements OAuth2User {
+
+    private final Long userId;
+    private final Collection<? extends GrantedAuthority> authorities;
+    private final Map<String, Object> attributes;
+    private final String nameAttributeKey;
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return attributes;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return authorities;
+    }
+
+    @Override
+    public String getName() {
+        return attributes.get(nameAttributeKey).toString();
+    }
+
+}
