+++ src/main/java/com/moonbaar/common/oauth/service/CustomOAuth2UserService.java
@@ -1,13 +1,14 @@
-package com.moonbaar.common.service;
+package com.moonbaar.common.oauth.service;
 
-import com.moonbaar.common.dto.OAuthAttributes;
+import com.moonbaar.common.oauth.CustomOAuth2User;
+import com.moonbaar.common.oauth.dto.OAuthAttributes;
+import com.moonbaar.domain.user.entity.User;
 import com.moonbaar.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 
@@ -33,10 +34,11 @@ public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequ
         OAuthAttributes attributes = OAuthAttributes.of(registrationId, userNameAttributeName, oAuth2User.getAttributes());
 
         // 회원가입 또는 프로필 업데이트 처리
-        userService.saveOrUpdate(attributes, registrationId);
+        User user = userService.saveOrUpdate(attributes, registrationId);
 
         // Spring Security가 인식할 사용자 객체 반환
-        return new DefaultOAuth2User(
+        return new CustomOAuth2User(
+                user.getId(),
                 Collections.singleton(() -> "ROLE_USER"),
                 attributes.attributes(),
                 attributes.userNameAttributeName()
